### 2017-09-05 #391 @whedon
[PRE REVIEW]: pyGPGO: Bayesian Optimization for Python
**Submitting author:** @hawk31 (<a href="http://orcid.org/0000-0002-5335-7834">Jiménez J.</a>)
**Repository:** <a href="https://github.com/hawk31/pyGPGO" target ="_blank">https://github.com/hawk31/pyGPGO</a>
**Version:** 0.3.0.dev1
