### 2017-09-10 71e3b6e
Create CONTRIBUTING.md
