+++ CONTRIBUTING.md
@@ -1,0 +1,32 @@
+First of all, thank you for wanting to contribute. It is people like you that render open-source software possible.
+Since pyGPGO is MIT Licensed, you can pretty much use the code here however it suits your needs, but if you want
+to contribute to the existing codebase, please consider the following guidelines.
+
+# How can I contribute to the project?
+
+Essentially either by:
+  1. Bug reporting.
+  2. Suggesting a feature
+  3. Contributing to the codebase
+  
+# Found a bug? Tell us about it!
+
+Hey, your bug may already be fixed! Try to update to the latest pyGPGO version
+(ideally the one provided in this repository) and check if it persists.
+
+Open an issue using the tag `[BUG]` and try to be as descriptive as possible. Also tell us about:
+  - your installed pyGPGO version.
+  - your Python environment setup.
+  - which OS you're running.
+
+Ideally, we would also like to have a minimal working example of the issue for reproducibility. 
+We'll fix your bug as soon as possible. 
+
+# Got a feature request? 
+
+We are always looking for interesting ideas to implement in the codebase. Open an issue with the tag
+`[REQUEST]` and describe your idea in detail, ideally providing some references.
+
+After appropriate discussion, if you're feeling adventurous, you can implement your own idea and open
+a pull request on the repository. We will review it in due time and approve it shall it meet our
+quality criteria.
