### 2017-09-15 5d4edec
Update paper.md
