### 2017-09-25 #391 @arfon
@openjournals/joss-editors - would anyone else be willing to edit this submission?
