### 2017-10-04 #391 @hawk31
It's been more than a month since I submitted this to JOSS. Does it usually take this long?
