### 2017-10-04 #391 @arfon
> It's been more than a month since I submitted this to JOSS. Does it usually take this long?

@hawk31 - we've got a pretty big backlog of submissions right now. We'll try and get to this soon.
