### 2017-10-04 #391 @labarba
Occasionally, when an editor is flagged but they are impossibly busy, we don't get a quick reply and there is a delay. But it is not usual. If you might suggest another editor, we'll try again.
