### 2017-10-16 #391 @arfon
@dfm - would you be willing to review this submission for JOSS?
