### 2017-10-16 2279e85
Update bib
