+++ paper.bib
@@ -88,3 +88,14 @@ url = {https://jmhldotorg.files.wordpress.com/2014/10/pes-final.pdf},
 year = {2014}
 }
 
+@article{Gardner2014,
+abstract = {Bayesian optimization is a powerful frame- work for minimizing expensive objective functions while using very few function eval- uations. It has been successfully applied to a variety of problems, including hyperparam- eter tuning and experimental design. How- ever, this framework has not been extended to the inequality-constrained optimization setting, particularly the setting in which eval- uating feasibility is just as expensive as eval- uating the objective. Here we present con- strained Bayesian optimization, which places a prior distribution on both the objective and the constraint functions. We evaluate our method on simulated and real data, demon- strating that constrained Bayesian optimiza- tion can quickly find optimal and feasible points, even when small feasible regions cause standard methods to fail.},
+author = {Gardner, Jacob R. and Kusner, Matt J. and Xu, Zhixiang Eddie and Weinberger, Kilian Q. and Cunningham, John P.},
+isbn = {9781634393973},
+journal = {Proceedings of the 31st International Conference on Machine Learning},
+pages = {937--945},
+title = {{Bayesian Optimization with Inequality Constraints}},
+volume = {32},
+year = {2014}
+}
+
