### 2017-10-16 #431 @whedon
[REVIEW]: pyGPGO: Bayesian Optimization for Python
**Submitting author:** @hawk31 (<a href="http://orcid.org/0000-0002-5335-7834">Jiménez J.</a>)
**Repository:** <a href="https://github.com/hawk31/pyGPGO" target ="_blank">https://github.com/hawk31/pyGPGO</a>
**Version:** 0.3.0.dev1

- [x] **Version:** Does the release version given match the GitHub release (0.3.0.dev1)?
