### 2017-10-16 #431 @dfm
@arfon: Does the PDF of the paper no longer get generated automatically?
