### 2017-10-17 d889304
Better unit and integration tests
