### 2017-10-17 e91e74e
Cleaning up MCMC classes
