### 2017-10-17 d3e8181
Fixing coverage
