+++ pyGPGO/surrogates/GaussianProcessMCMC.py
@@ -1,4 +1,3 @@
-# work in progress
 import numpy as np
 import scipy as sp
 import theano
