### 2017-10-17 6b8e9e6
Cleaning up
