+++ build/lib/pyGPGO/surrogates/tStudentProcess.py
@@ -1,245 +1,0 @@
-import numpy as np
-from collections import OrderedDict
-from numpy.linalg import slogdet
-from scipy.linalg import inv
-from scipy.optimize import minimize
-from scipy.special import gamma
-
-def logpdf(x, df, mu, Sigma):
-    """
-    Marginal log-likelihood of a Student-t Process
-    
-    Parameters
-    ----------
-    x: array-like
-        Point to be evaluated
-    df: float
-        Degrees of freedom (>2.0)
-    mu: array-like
-        Mean of the process.
-    Sigma: array-like
-        Covariance matrix of the process.
-
-    Returns
-    -------
-    logp: float
-        log-likelihood 
-
-    """
-    d = len(x)
-    x = np.atleast_2d(x)
-    xm = x - mu
-    V = df * Sigma
-    V_inv = np.linalg.inv(V)
-    _, logdet = slogdet(np.pi * V)
-
-    logz = -gamma(df / 2.0 + d / 2.0) + gamma(df / 2.0) + 0.5 * logdet
-    logp = -0.5 * (df + d) * np.log(1 + np.sum(np.dot(xm, V_inv) * xm, axis=1))
-
-    logp = logp - logz
-
-    return logp[0]
-
-
-class tStudentProcess:
-    def __init__(self, covfunc, nu=3.0, optimize=False):
-        """
-        t-Student Process regressor class. This class DOES NOT support gradients in ML estimation yet.
-        
-        Parameters
-        ----------
-        covfunc: instance from a class of covfunc module
-            An instance from a class from the `covfunc` module.
-        nu: float
-            (>2.0) Degrees of freedom
-
-        Attributes
-        ----------
-        covfunc: object
-            Internal covariance function.
-        nu: float
-            Degrees of freedom.
-        optimize: bool
-            Whether to optimize covariance function hyperparameters.
-
-        """
-        self.covfunc = covfunc
-        self.nu = nu
-        self.optimize = optimize
-
-    def getcovparams(self):
-        """
-        Returns current covariance function hyperparameters
-
-        Returns
-        -------
-        dict
-            Dictionary containing covariance function hyperparameters
-        """
-        d = {}
-        for param in self.covfunc.parameters:
-            d[param] = self.covfunc.__dict__[param]
-        return d
-
-    def _lmlik(self, param_vector, param_key):
-        """
-        Returns marginal negative log-likelihood for given covariance hyperparameters.
-
-        Parameters
-        ----------
-        param_vector: list
-            List of values corresponding to hyperparameters to query.
-        param_key: list
-            List of hyperparameter strings corresponding to `param_vector`.
-
-        Returns
-        -------
-        float
-            Negative log-marginal likelihood for chosen hyperparameters.
-
-        """
-        k_param = OrderedDict()
-        for k, v in zip(param_key, param_vector):
-            k_param[k] = v
-        self.covfunc = self.covfunc.__class__(**k_param)
-
-        # This fixes recursion
-        original_opt = self.optimize
-        self.optimize = False
-        self.fit(self.X, self.y)
-        self.optimize = original_opt
-
-        return (- self.logp)
-
-    def optHyp(self, param_key, param_bounds, n_trials=5):
-        """
-        Optimizes the negative marginal log-likelihood for given hyperparameters and bounds.
-        This is an empirical Bayes approach (or Type II maximum-likelihood).
-
-        Parameters
-        ----------
-        param_key: list
[…]
