### 2017-10-17 d3e2a5d
Reduced testing time
