### 2017-10-17 5333692
Reduced testing time
