+++ .travis.yml
@@ -9,7 +9,6 @@ install:
   - "pip install ."
   - "pip install pytest-cov"
   - "pip install codecov"
-  - travis_wait 30 mvn install
 # command to run tests
 script: py.test --cov=./
 # codecov
