### 2017-10-17 fec1c8d
Coverage only on test suite
