### 2017-10-17 c945cdd
cleaning unit tests
