+++ pyGPGO.egg-info/SOURCES.txt
@@ -1,22 +1,0 @@
-LICENSE
-MANIFEST.in
-setup.py
-pyGPGO/GPGO.py
-pyGPGO/__init__.py
-pyGPGO/acquisition.py
-pyGPGO/covfunc.py
-pyGPGO/logger.py
-pyGPGO/version.py
-pyGPGO.egg-info/PKG-INFO
-pyGPGO.egg-info/SOURCES.txt
-pyGPGO.egg-info/dependency_links.txt
-pyGPGO.egg-info/not-zip-safe
-pyGPGO.egg-info/requires.txt
-pyGPGO.egg-info/top_level.txt
-pyGPGO/surrogates/BoostedTrees.py
-pyGPGO/surrogates/GaussianProcess.py
-pyGPGO/surrogates/GaussianProcessMCMC.py
-pyGPGO/surrogates/RandomForest.py
-pyGPGO/surrogates/__init__.py
-pyGPGO/surrogates/tStudentProcess.py
-pyGPGO/surrogates/tStudentProcessMCMC.py
