### 2017-10-17 d93e4d9
cleaned logger
