+++ pyGPGO/logger.py
@@ -28,24 +28,3 @@ class EventLogger:
         for init_eval in range(gpgo.init_evals):
             print(self.template.format('init', gpgo.GP.X[init_eval], gpgo.GP.y[init_eval], gpgo.tau))
 
-if __name__ == '__main__':
-    import numpy as np
-    import matplotlib.pyplot as plt
-    from pyGPGO.covfunc import squaredExponential
-    from pyGPGO.surrogates.GaussianProcess import GaussianProcess
-    from pyGPGO.acquisition import Acquisition
-    from pyGPGO.GPGO import GPGO
-
-    np.random.seed(20)
-
-    def f(x):
-        return -((6*x-2)**2*np.sin(12*x-4))
-
-    sexp = squaredExponential()
-    gp = GaussianProcess(sexp)
-    acq = Acquisition(mode = 'ExpectedImprovement')
-
-    params = {'x': ('cont', (0, 1))}
-    gpgo = GPGO(gp, acq, f, params)
-    gpgo.run(max_iter = 10)
-    print(gpgo.getResult())
