### 2017-10-18 f124860
Added tutorials and commented examples
