+++ examples/minimalexample.py
@@ -1,3 +1,8 @@
+#######################################
+# pyGPGO examples
+# minimalexample: A minimal working pyGPGO example.
+#######################################
+
 import numpy as np
 import matplotlib.pyplot as plt
 from pyGPGO.covfunc import squaredExponential
@@ -5,12 +10,14 @@ from pyGPGO.surrogates.GaussianProcess import GaussianProcess
 from pyGPGO.acquisition import Acquisition
 from pyGPGO.GPGO import GPGO
 
+
 def drawFun(f):
     x = np.linspace(0, 1, 1000)
     plt.plot(x, f(x))
     plt.grid()
     plt.show()
-    
+
+
 if __name__ == '__main__':
     np.random.seed(20)
     def f(x):
@@ -26,4 +33,3 @@ if __name__ == '__main__':
     gpgo = GPGO(gp, acq, f, params)
     gpgo.run(max_iter = 10)
     print(gpgo.getResult())
-
