### 2017-10-18 c19e5ea
modified docs source
