+++ docs/build/html/_sources/pyGPGO.rst.txt
@@ -1,0 +1,28 @@
+pyGPGO package
+==============
+
+Subpackages
+-----------
+
+.. toctree::
+
+    pyGPGO.surrogates
+
+Submodules
+----------
+
+.. toctree::
+
+   pyGPGO.GPGO
+   pyGPGO.acquisition
+   pyGPGO.covfunc
+   pyGPGO.logger
+   pyGPGO.version
+
+Module contents
+---------------
+
+.. automodule:: pyGPGO
+    :members:
+    :undoc-members:
+    :show-inheritance:
