+++ docs/build/html/_sources/api.rst.txt
@@ -1,0 +1,14 @@
+pyGPGO documentation
+==================
+
+
+Contents:
+
+.. toctree::
+    :maxdepth: 3
+
+    Bayesian Optimization module <pyGPGO.GPGO>
+    Surrogates module <pyGPGO.surrogates>
+    Covariance function module <pyGPGO.covfunc>
+    Acquisition function module <pyGPGO.acquisition>
+
