### 2017-10-18 4e355c3
mocking h5py
