+++ requirements_rtd.txt
@@ -1,0 +1,5 @@
+numpy
+scipy
+joblib
+scikit-learn
+pymc3
