### 2017-10-18 ff98249
mocking theano
