### 2017-10-18 b3b6752
unmocking theano
