+++ docs/source/conf.py
@@ -367,5 +367,5 @@ class Mock(MagicMock):
     def __getattr__(cls, name):
             return MagicMock()
 
-MOCK_MODULES = ['h5py', 'theano']
+MOCK_MODULES = ['h5py']
 sys.modules.update((mod_name, Mock()) for mod_name in MOCK_MODULES)
