+++ requirements_rtd.txt
@@ -3,3 +3,4 @@ scipy
 joblib
 scikit-learn
 pymc3
+theano
