### 2017-10-19 878ec13
Updated docs with comparison and features
