### 2017-10-19 7a7c71e
Missing reference in index.rst
