### 2017-10-21 d11dca7
Updated doc links
