### 2017-10-23 35172be
Fixed citations in docs
