### 2017-10-23 05ffaaf
fixing citations in docs
