### 2017-10-23 b53f4d5
fixing citations in docs
