### 2017-10-23 7753ef5
fixing citations in docs
