### 2017-10-30 #431 @dfm
We're getting close here, but I have three remaining issues:

1. The installation instructions should discuss the required and optional dependencies (e.g. pymc3).
2. The statement of need in the paper is pretty weak. Right now, there isn't any explicit discussion of why Bayesian optimization is preferred over other methods or why pyGPGO is needed when there are other existing packages.
3. The final sentence of the first paragraph of the paper is a description of Gaussian Processes (and perhaps, if we're generous, Bayesian inference in general), but it is not really relevant to Bayesian optimization. As far as I can tell, Bayesian optimization is used because it exploits "smoothness" assumptions to minimize the number of function evaluations.
