+++ paper.bib
@@ -99,3 +99,38 @@ volume = {32},
 year = {2014}
 }
 
+@article{Chapelle2011,
+abstract = {Thompson sampling is one of oldest heuristic to address the exploration ex- ploitation trade-off, but it is surprisingly unpopular in the literature. We present here some empirical results using Thompson sampling on simulated and real data, and show that it is highly competitive. And since this heuristic is very easy to implement, we argue that it should be part of the standard baselines to compare against.},
+author = {Chapelle, Olivier and Li, Lihong},
+isbn = {9781618395993},
+journal = {Advances in Neural Information Processing Systems},
+pages = {2249----2257},
+title = {{An Empirical Evaluation of Thompson Sampling}},
+url = {http://explo.cs.ucl.ac.uk/wp-content/uploads/2011/05/An-Empirical-Evaluation-of-Thompson-Sampling-Chapelle-Li-2011.pdf},
+year = {2011}
+}
+
+@inproceedings{Vanchinathan2014,
+abstract = {We address the challenge of ranking recommendation lists based on click feedback by efficiently encoding similarities among users and among items. The key challenges are threefold: (1) combinato- rial number of lists; (2) sparse feedback and (3) context dependent recommendations. We propose the CGPRANK algorithm, which exploits prior information specified in terms of a Gaussian pro- cess kernel function, which allows to share feedback in three ways: Between positions in a list, between items, and between contexts. Under our model, we provide strong performance guarantees and empirically evaluate our algorithm on data from two large scale recommendation tasks: Yahoo! news article recommendation, and Google books. In our experiments, CGPRANK significantly out- performs state-of-the-art multi-armed bandit and learning-to-rank methods, with an 18{\%} increase in clicks.},
+author = {Vanchinathan, Hastagiri P. and Nikolic, Isidor and {De Bona}, Fabio and Krause, Andreas},
+booktitle = {Proceedings of the 8th ACM Conference on Recommender systems - RecSys '14},
+doi = {10.1145/2645710.2645733},
+isbn = {9781450326681},
+pages = {225--232},
+title = {{Explore-exploit in top-N recommender systems via Gaussian processes}},
+url = {http://dl.acm.org/citation.cfm?doid=2645710.2645733},
+year = {2014}
+}
+
+
+@article{Salvatier2016,
+abstract = {Probabilistic programming allows for automatic Bayesian inference on user-defined probabilistic models. Recent advances in Markov chain Monte Carlo (MCMC) sampling allow inference on increasingly complex models. This class of MCMC, known as Hamiltonian Monte Carlo, requires gradient information which is often not readily available. PyMC3 is a new open source probabilistic programming framework written in Python that uses Theano to compute gradients via automatic differentiation as well as compile probabilistic programs on-the-fly to C for increased speed. Contrary to other probabilistic programming languages, PyMC3 allows model specification directly in Python code. The lack of a domain specific language allows for great flexibility and direct interaction with the model. This paper is a tutorial-style introduction to this software package.},
+author = {Salvatier, J and Wiecki, TV and Fonnesbeck C.},
+journal = {PeerJ Computer Science},
+title = {{Probabilistic programming in Python using PyMC3}},
+url = {https://doi.org/10.7717/peerj-cs.55},
+year = {2016}
+}
+
+
+
