### 2017-10-30 d830ad8
Update paper.md
