### 2017-10-30 8f758b0
Update paper.md
