### 2017-10-30 e377388
Update paper.md
