### 2017-10-30 #431 @hawk31
Hi @dfm @arfon !

I'm currently addressing these issues:

1. Both `pyMC3` and `theano` are now both _required_ dependencies for `pyGPGO`. While most of the package functionality is not affected by these dependencies (that is, everything that is not MCMC sampling dependent) the installation procedure is significantly simplified to a simple `pip install pyGPGO`.

2. I've updated the `paper.md` and `paper.bib` to stress both pyGPGO's and the framework's statement of need. A feature-wise comparison between pyGPGO and other software for Bayesian Optimization is available in the [software documentation](http://pygpgo.readthedocs.io/en/latest/comparison.html).

Hopefully the paper should be significantly better written now. Let me know if anything else needs to be addressed.
