### 2017-10-30 #431 @dfm
Cool. Much better! It might be worth *listing* these dependancies somewhere so that users know what it is being used under the hood.
