### 2017-11-01 0db7ffe
Updated docs with dependency list
