### 2017-11-01 a7774ec
Update paper.md
