### 2017-11-01 #431 @hawk31
Hi @dfm !

Dependencies are now listed on the GitHub page of the project, on the documentation and on the `paper.md` file.

Again let me know if you would like to have something improved :)
