### 2017-11-01 #431 @arfon
```
pandoc-citeproc: reference SpearmingSnoek2012 not found
```

@hawk31 - looks like there's a slight issue with one of the citations ☝️. Once you've fixed that, could you make an archive of the reviewed software in Zenodo/figshare/other service and update this thread with the DOI of the archive? I can then move forward with accepting the submission.
