### 2017-11-02 c1482fd
Fixed citation typo in paper.md
