### 2017-11-02 #431 @hawk31
Hi @arfon !

Whoops, I fixed that typo already on the master branch and made a new release with all the changes. The zenodo archive of the software is available in https://zenodo.org/record/1040676

Thanks!
